fixed dto types to be api-compatible

+++ app/src/main/java/ru/practicum/android/diploma/data/dto/VacanciesResponse.kt
@@ -1,3 +1,6 @@
 package ru.practicum.android.diploma.data.dto
 
-data class VacanciesResponse(val found: Long, val pages: Long, val page: Long, val items: List<VacancyDto>) : Response()
+data class VacanciesResponse(val found: Int,
+                             val pages: Int,
+                             val page: Int,
+                             val items: List<VacancyDto>) : Response()
